+++ CHANGELOG.md
@@ -15,8 +15,8 @@ v2.0.0-beta.10
  * ons-icon: Delete attribute `flip` from ons-icon (obsolete, since everything can be done with `rotate`)
  * ons-list-item: Make the attribute `tappable` a boolean and add attribute `tap-background-color`
  * ons-alert-dialog: Fixed [#1347](https://github.com/OnsenUI/OnsenUI/issues/1347).
- * ons-progress-bar: Add `value`, `secondaryValue` and `indeterminate` properties.
- * ons-progress-circular: Add `value`, `secondaryValue` and `indeterminate` properties.
+ * ons-alert-dialog, ons-dialog, ons-popover, ons-modal: Updated API (`disabled`, `visible`, `backButtonHandler` properties).
+ * ons-progress-bar, ons-progress-circular: Add `value`, `secondaryValue` and `indeterminate` properties.
  * ons-pull-hook: Add `height`, `thresholdHeight` and `disabled` properties.
 
 v2.0.0-beta.9